Extracted all things from CKEDITOR object to avoid circ deps.
